feat(manage-users): refactor user management components and add new dropdowns for location and multi-select
