+++ src/features/manage-users/components/index.ts
@@ -1,2 +1,8 @@
+export * from "./delete-user-dialog";
 export * from "./edit-user-dialog";
+export * from "./location-search-dropdown";
 export * from "./manage-users-page";
+export * from "./manage-users-pagination";
+export * from "./manage-users-table";
+export * from "./manage-users-toolbar";
+export * from "./multi-select-dropdown";
