+++ src/features/manage-users/components/manage-users-page.tsx
@@ -1,55 +1,16 @@
 "use client";
 
-import { ArrowDownUp, Download, Pencil, Search, Trash2 } from "lucide-react";
 import { useEffect, useMemo, useState } from "react";
 import { toast } from "sonner";
 import { endpoints } from "@/api/endpoints";
-import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Spinner } from "@/components/ui/spinner";
 import { authStore } from "@/lib/auth";
 import { useDeleteManageUser, useManageUsersList } from "../hooks";
+import { DeleteUserDialog } from "./delete-user-dialog";
 import { EditUserDialog } from "./edit-user-dialog";
-
-const columnOrder = [
-  { key: "full_name", label: "Full Name", sortable: true },
-  { key: "karma", label: "Total Karma", sortable: true },
-  { key: "muid", label: "Mu ID", sortable: true },
-  { key: "email", label: "Email", sortable: true },
-  { key: "mobile", label: "Mobile", sortable: true },
-  { key: "discord_id", label: "Discord ID", sortable: true },
-  { key: "level", label: "Level", sortable: true },
-  { key: "created_at", label: "Created On", sortable: true },
-] as const;
-
-function formatDate(value: string | null | undefined) {
-  if (!value) return "-";
-
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return value;
-
-  return date.toLocaleDateString("en-US", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  });
-}
+import { ManageUsersPagination } from "./manage-users-pagination";
+import { ManageUsersTable } from "./manage-users-table";
+import { ManageUsersToolbar } from "./manage-users-toolbar";
 
 function getCsvUrl() {
   const base = process.env.NEXT_PUBLIC_DJANGO_API_URL;
@@ -172,187 +133,40 @@ export function ManageUsersPage() {
           <CardTitle className="text-xl text-foreground">
             Manage Users
           </CardTitle>
-
-          <div className="flex w-full flex-col gap-2 sm:w-auto sm:flex-row">
-            <div className="relative min-w-[260px]">
-              <Search className="pointer-events-none absolute left-3 top-1/2 size-4 -translate-y-1/2 text-muted-foreground" />
-              <Input
-                value={searchInput}
-                onChange={(event) => setSearchInput(event.target.value)}
-                placeholder="Search users"
-                className="pl-9"
-              />
-            </div>
-
-            <Select
-              value={String(perPage)}
-              onValueChange={(value) => {
-                setPageIndex(1);
-                setPerPage(Number(value));
-              }}
-            >
-              <SelectTrigger className="w-full sm:w-[120px]">
-                <SelectValue placeholder="Rows" />
-              </SelectTrigger>
-              <SelectContent>
-                {[5, 10, 20, 50, 100].map((option) => (
-                  <SelectItem key={option} value={String(option)}>
-                    {option} / page
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-
-            <Button variant="outline" onClick={handleDownloadCsv}>
-              <Download className="size-4" />
-              CSV
-            </Button>
-          </div>
+          <ManageUsersToolbar
+            searchInput={searchInput}
+            perPage={perPage}
+            onSearchChange={setSearchInput}
+            onPerPageChange={(value) => {
+              setPageIndex(1);
+              setPerPage(value);
+            }}
+            onDownloadCsv={handleDownloadCsv}
+          />
         </CardHeader>
 
         <CardContent className="space-y-4">
-          <div className="overflow-x-auto rounded-lg border border-border">
-            <table className="w-full min-w-[980px] border-collapse">
-              <thead className="bg-muted/50">
-                <tr>
-                  <th className="px-4 py-3 text-left text-sm font-semibold text-foreground">
-                    Sl.no
-                  </th>
+          <ManageUsersTable
+            users={users}
+            isLoading={isLoading}
+            pageIndex={pageIndex}
+            perPage={perPage}
+            onSort={toggleSort}
+            onEdit={openEdit}
+            onDelete={openDelete}
+          />
 
-                  {columnOrder.map((column) => (
-                    <th
-                      key={column.key}
-                      className="px-4 py-3 text-left text-sm font-semibold text-foreground"
-                    >
-                      <button
-                        type="button"
-                        className="inline-flex items-center gap-1"
-                        onClick={() =>
-                          column.sortable && toggleSort(column.key)
-                        }
-                      >
-                        <span>{column.label}</span>
-                        {column.sortable && (
-                          <ArrowDownUp className="size-4 text-muted-foreground" />
-                        )}
-                      </button>
-                    </th>
-                  ))}
-
-                  <th className="px-4 py-3 text-left text-sm font-semibold text-foreground">
-                    Action
-                  </th>
-                </tr>
-              </thead>
-
-              <tbody>
-                {isLoading ? (
-                  <tr>
-                    <td colSpan={columnOrder.length + 2} className="px-4 py-10">
-                      <div className="flex items-center justify-center">
-                        <Spinner className="size-6 text-primary" />
-                      </div>
-                    </td>
-                  </tr>
-                ) : users.length === 0 ? (
-                  <tr>
-                    <td
-                      colSpan={columnOrder.length + 2}
-                      className="px-4 py-10 text-center text-sm text-muted-foreground"
-                    >
-                      No data to display
-                    </td>
-                  </tr>
-                ) : (
-                  users.map((user, index) => (
-                    <tr key={user.id} className="border-t border-border">
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {(pageIndex - 1) * perPage + index + 1}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {user.full_name || "-"}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {user.karma}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {user.muid || "-"}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {user.email || "-"}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {user.mobile || "-"}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {user.discord_id || "-"}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {user.level}
-                      </td>
-                      <td className="px-4 py-3 text-sm text-foreground">
-                        {formatDate(user.created_at)}
-                      </td>
-                      <td className="px-4 py-3">
-                        <div className="flex items-center gap-2">
-                          <Button
-                            type="button"
-                            variant="ghost"
-                            size="icon-sm"
-                            onClick={() => openEdit(user.id)}
-                            aria-label="Edit user"
-                          >
-                            <Pencil className="size-4" />
-                          </Button>
-                          <Button
-                            type="button"
-                            variant="ghost"
-                            size="icon-sm"
-                            onClick={() => openDelete(user.id)}
-                            aria-label="Delete user"
-                          >
-                            <Trash2 className="size-4 text-destructive" />
-                          </Button>
-                        </div>
-                      </td>
-                    </tr>
-                  ))
-                )}
-              </tbody>
-            </table>
-          </div>
-
-          <div className="flex flex-col gap-3 rounded-lg border border-border bg-card p-3 sm:flex-row sm:items-center sm:justify-between">
-            <p className="text-sm text-muted-foreground">
-              {startItem} - {endItem}
-            </p>
-
-            <div className="flex items-center gap-2">
-              <Button
-                type="button"
-                variant="outline"
-                onClick={() =>
-                  setPageIndex((current) => Math.max(1, current - 1))
-                }
-                disabled={pageIndex <= 1 || isFetching}
-              >
-                Previous
-              </Button>
-              <span className="min-w-20 text-center text-sm text-foreground">
-                {pageIndex} / {Math.max(totalPages, 1)}
-              </span>
-              <Button
-                type="button"
-                variant="outline"
-                onClick={() => setPageIndex((current) => current + 1)}
-                disabled={
-                  pageIndex >= totalPages || totalPages === 0 || isFetching
-                }
-              >
-                Next
-              </Button>
-            </div>
-          </div>
+          <ManageUsersPagination
+            startItem={startItem}
+            endItem={endItem}
+            pageIndex={pageIndex}
+            totalPages={totalPages}
+            isFetching={isFetching}
+            onPrevious={() =>
+              setPageIndex((current) => Math.max(1, current - 1))
+            }
+            onNext={() => setPageIndex((current) => current + 1)}
+          />
         </CardContent>
       </Card>
 
@@ -362,35 +176,12 @@ export function ManageUsersPage() {
         userId={editingUserId}
       />
 
-      <Dialog open={isDeleteOpen} onOpenChange={setIsDeleteOpen}>
-        <DialogContent className="sm:max-w-md">
-          <DialogHeader>
-            <DialogTitle>Delete User</DialogTitle>
-            <DialogDescription>
-              Are you sure you want to delete this user?
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => setIsDeleteOpen(false)}
-              disabled={deleteMutation.isPending}
-            >
-              Cancel
-            </Button>
-            <Button
-              type="button"
-              variant="destructive"
-              onClick={handleDelete}
-              disabled={deleteMutation.isPending}
-            >
-              {deleteMutation.isPending && <Spinner className="mr-2" />}
-              Delete
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <DeleteUserDialog
+        open={isDeleteOpen}
+        isDeleting={deleteMutation.isPending}
+        onOpenChange={setIsDeleteOpen}
+        onConfirm={handleDelete}
+      />
     </div>
   );
 }
